Fix cannot add view to mFrameLayout when extends Cocos2dxActivity
