+++ cocos/platform/android/java/src/org/cocos2dx/lib/Cocos2dxActivity.java
@@ -50,46 +50,6 @@ import javax.microedition.khronos.egl.EGL10;
 import javax.microedition.khronos.egl.EGLConfig;
 import javax.microedition.khronos.egl.EGLDisplay;
 
-class ResizeLayout extends FrameLayout{
-    private  boolean mEnableForceDoLayout = false;
-
-    public ResizeLayout(Context context){
-        super(context);
-    }
-
-    public ResizeLayout(Context context, AttributeSet attrs) {
-        super(context, attrs);
-    }
-
-    public void setEnableForceDoLayout(boolean flag){
-        mEnableForceDoLayout = flag;
-    }
-
-    @Override
-    protected void onLayout(boolean changed, int l, int t, int r, int b) {
-        super.onLayout(changed, l, t, r, b);
-        if(mEnableForceDoLayout){
-            /*This is a hot-fix for some android devices which don't do layout when the main window
-            * is paned.  We refersh the layout in 24 frames per seconds.
-            * When the editBox is lose focus or when user begin to type, the do layout is disabled.
-            */
-            final Handler handler = new Handler();
-            handler.postDelayed(new Runnable() {
-                @Override
-                public void run() {
-                    //Do something after 100ms
-                    requestLayout();
-                    invalidate();
-                }
-            }, 1000 / 24);
-
-        }
-
-    }
-
-}
-
-
 public abstract class Cocos2dxActivity extends Activity implements Cocos2dxHelperListener {
     // ===========================================================
     // Constants
